chore: refactor block caches
